calendar-tools/requirements.txt
google-api-python-client>=2.125.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
matplotlib>=3.8.0
